Split DefaultExceptionHandler into strict and forgiving versions
Currently DefaultExceptionHandler will close the channel
when an unhandled consumer or listener exception occurs.
Not everybody agrees this is a great idea. This commit introduces
a forgiving version and keeps DefaultExceptionHandler strict.

For 3.6.0 we will make DefaultExceptionHandler forgiving.
